requirements.txt
# Core dependencies for BabyVis - Compatible versions tested on RTX 3050 Ti
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
diffusers==0.18.0
transformers==4.34.0
accelerate==0.20.3

# Image processing
Pillow>=9.0.0
opencv-python>=4.7.0

# Scientific computing
numpy<2
scipy>=1.9.0

# Utilities
huggingface-hub==0.16.4
safetensors>=0.3.0
psutil>=5.9.0

# Optional: for web interface
gradio>=3.40.0

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Note: xformers is intentionally excluded due to compatibility issues with this PyTorch version
